removed other modes, add build version, added screenshots, remove unused api calls
