+++ components/select-station.component.jsx
@@ -2,6 +2,7 @@ import {ActivityIndicator, Text, TouchableOpacity, View} from "react-native";
 import {useContext} from "react";
 import {StationContext} from "../contexts/station.context";
 import Banner from "./banner.component";
+import Version from "./version.component";
 
 const SelectStation = () => {
   const { stations, setStationId, loading } = useContext(StationContext)
@@ -12,39 +13,44 @@ const SelectStation = () => {
 
   return (
     <View style={{
-      flex: 1,
-      alignItems: 'center',
-      justifyContent: 'center'
+      flex: 1
     }}>
-      <Banner />
+      <View style={{
+        flex: 1,
+        alignItems: 'center',
+        justifyContent: 'center'
+      }}>
+        <Banner />
 
-      {
-        loading
-          ?
-            <>
-              <ActivityIndicator />
-              <Text>Loading stations...</Text>
-            </>
-          : (
-            <>
-            <Text style={{
-              fontSize: 24,
-              color: '#1434a3'
-            }}>SELECT STATION</Text>
-              {
-                stations.map(({ name, id }) => (
-                <TouchableOpacity style={{
-                marginTop: 10
-              }} key={id} onPress={() => handleSetStation(id)}>
-                <Text>{name}</Text>
-                </TouchableOpacity>
-                ))
-              }
-            </>
-          )
-      }
+        {
+          loading
+            ?
+              <>
+                <ActivityIndicator />
+                <Text>Loading stations...</Text>
+              </>
+            : (
+              <>
+              <Text style={{
+                fontSize: 24,
+                color: '#1434a3'
+              }}>SELECT STATION</Text>
+                {
+                  stations.map(({ name, id }) => (
+                  <TouchableOpacity style={{
+                  marginTop: 10
+                }} key={id} onPress={() => handleSetStation(id)}>
+                  <Text>{name}</Text>
+                  </TouchableOpacity>
+                  ))
+                }
+              </>
+            )
+        }
+      </View>
+
+      <Version />
     </View>
-
   )
 }
