+++ utils.js
@@ -106,76 +106,6 @@ export const submitQrScans = async (scansToSubmit) => {
     }
 }
 
-/**
- * Handle submission of faces data to backend
- *
- * @returns {Promise<*>}
- * @param data
- */
-export const submitCount = async (data) => {
-    console.log(`scansToSubmit: `, data)
-
-    try {
-        const response = await api.post('/passenger-count', data)
-
-        const { message, success } = response.data
-
-        showToastMessage(message)
-
-        return success
-    } catch (e) {
-        console.log(`error in submitCount: `,e)
-
-        showToastMessage(e.message)
-    }
-}
-
-/**
- * Handle submission of faces data to backend
- *
- * @returns {Promise<*>}
- * @param passengers
- */
-export const submitPassengers = async (passengers) => {
-    try {
-        const response = await api.post('/kiosks', {
-            passengers: passengers
-        })
-
-        const { success, message } = response.data
-
-        showToastMessage(message)
-
-        return success
-    } catch (e) {
-        console.log(`error in submitPassengers: `, e)
-
-        showToastMessage(e.message)
-    }
-}
-
-/**
- * Handle submission of faces data to backend
- *
- * @param facesToSubmit
- * @returns {Promise<*>}
- */
-export const submitFaces = async (facesToSubmit) => {
-    try {
-        const response = await api.post('/faces', {
-            faces: facesToSubmit
-        })
-
-        const { message, success } = response.data
-
-        showToastMessage(message)
-
-        return success
-    } catch (e) {
-        console.log(`error in submitFaces: `, e.message)
-    }
-}
-
 const showToastMessage = (message) => {
     Toast.show(message, {
         duration: Toast.durations.SHORT,
